Move playwright signIn logic to test fixtures
